Formulario para modificar e-mail agregado
