+++ src/app/(app)/profile/layout.js
@@ -1,18 +1,14 @@
 import Sidebar from '@/components/layout/Sidebar'
 import { Col, Container, Row } from 'react-bootstrap'
 
-export const metadata = {
-  title: 'My profile',
-}
-
 export default function MyProfileLayout({ children }) {
   return (
-    <Row className='g-0'>
-      <Col xs={2}>
+    <Row className="g-0 sidebar-container">
+      <Col className="col-auto d-flex">
         <Sidebar />
       </Col>
-      <Col xs={10}>
-        <Container className="mt-5">{children}</Container>
+      <Col>
+        <Container className="my-5 px-5">{children}</Container>
       </Col>
     </Row>
   )
