+++ src/app/(app)/profile/modify-data/page.js
@@ -1,0 +1,9 @@
+import DataForm from './DataForm'
+
+export const metadata = {
+  title: 'Corrección de datos',
+}
+
+export default function DataPage() {
+  return <DataForm />
+}
