+++ src/app/(app)/profile/modify-display-name/page.js
@@ -1,0 +1,9 @@
+import DisplayNameForm from './DisplayNameForm'
+
+export const metadata = {
+  title: '¿Cómo quieres que te llamemos?',
+}
+
+export default function DisplayNamePage() {
+  return <DisplayNameForm />
+}
