+++ src/app/(app)/profile/modify-email/page.js
@@ -1,0 +1,9 @@
+import EmailForm from './EmailForm'
+
+export const metadata = {
+  title: 'Ingresa tu e-mail',
+}
+
+export default function EmailPage() {
+  return <EmailForm />
+}
