+++ src/app/(app)/profile/modify-phone/page.js
@@ -1,0 +1,9 @@
+import PhoneForm from './PhoneForm'
+
+export const metadata = {
+  title: 'Ingresa tu número de celular',
+}
+
+export default function PhonePage() {
+  return <PhoneForm />
+}
