--- src/app/(app)/profile/personal-information/layout.js
@@ -1,7 +1,0 @@
-export const metadata = {
-  title: 'Personal information',
-}
-
-export default function PersonalInformationLayout({ children }) {
-  return <>{children}</>
-}
